requirements.txt
alembic==1.8.0
bidict==0.22.0
click==8.1.3
dnspython==2.2.1
dominate==2.6.0
eventlet==0.33.1
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-MQTT==1.1.1
Flask-Script==2.0.6
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
paho-mqtt==1.6.1
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.3
python-socketio==5.7.0
six==1.16.0
SQLAlchemy==1.4.39
visitor==0.1.3
Werkzeug==2.1.2
zipp==3.8.1
